requirements.txt
streamlit
pandas
numpy
scikit-learn
transformers
torch
faiss-cpu